configure routers
