+++ express/index.js
@@ -1,0 +1,4 @@
+const talkerRouter = require('./talkerRouter');
+const loginRouter = require('./loginRouter');
+
+module.exports = { talkerRouter, loginRouter };
